+++ CHANGELOG.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Added
-* ...
+* [extras/docker] docker engine (not docker desktop)
 
 ### Changed
 * [extras/vscode] a first list of extensions